+++ src/estimation/include/iDynTree/Estimation/ExtWrenchesAndJointTorquesEstimator.h
@@ -105,20 +105,6 @@ public:
     bool setModelAndSensors(const Model & _model, const SensorsList & _sensors);
 
     /**
-     * Load model and sensors from file.
-     *
-     * @deprecated Use iDynTree::ModelLoader::loadModelFromFile and call setModelAndSensors
-     * on the parsed Model and SensorsList
-     *
-     * @param[in] filename path to the file to load.
-     * @param[in] filetype (optional) explicit definiton of the filetype to load.
-     *                     Only "urdf" is supported at the moment.
-     * @return true if all went well (files were correctly loaded and consistent), false otherwise.
-     */
-    IDYNTREE_DEPRECATED_WITH_MSG("Use iDynTree::ModelLoader::loadModelFromFile and call setModelAndSensors on the parsed Model and SensorsList")
-    bool loadModelAndSensorsFromFile(const std::string filename, const std::string filetype="");
-
-    /**
      * Load model and sensors from file, specifieng the dof considered for the estimation.
      *
      * @note this will create e a reduced model only with the joint specified in consideredDOFs and  the